[OTAWA] has a 'sporadic' bug in PERS analysis. Here is a script that checks whether analysis was bug-free
If a BB is very long, then its tail may evict its head. OTAWA however
fails to detect this (known bug, Roychoudhury et al. found it in 2013
in theory). As a consequence, WCET can be underestimated.

This script simply checks when blocks are so long that they wrap into
the cache set more often than associativity, and then checks that the
l-blocks subject to definitive eviciton are not classified anything else
than miss.